backend/calorie_backend/requirements.txt
Django==4.2.27
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-decouple==3.8
pymysql==1.1.2
bcrypt==5.0.0
Pillow==11.1.0
boto3==1.36.26
markdown==3.7
bleach==6.2.0
PyJWT==2.10.1
